django_cn/requirements.txt
Django==1.6.2
MySQL-python==1.2.5
Pillow==2.4.0
PyPDF2==1.21
South==0.8.4
argparse==1.2.1
django-auth-ldap==1.2.0
django-cleanup==0.1.10
django-debug-toolbar==1.2.1
django-easy-pdf==0.1.0
django-extensions==1.3.7
django-filebrowser==3.5.6
django-grappelli==2.5.3
django-markdown-deux==1.0.4
html5lib==0.999
markdown2==2.2.1
netaddr==0.7.11
pisa==3.0.33
pyPdf==1.13
pydot==1.0.2
pyparsing==1.5.7
python-ldap==2.4.15
pytz==2014.3
reportlab==2.7
six==1.6.1
sqlparse==0.1.11
wsgiref==0.1.2
xhtml2pdf==0.0.6
